Use uintptr_t and intptr_t for pointer-integer casts

+++ libs/JSystem/include/JSystem/JKernel/JKRAramPiece.hpp
@@ -1,6 +1,7 @@
 #ifndef JKR_ARAM_PIECE_HPP
 #define JKR_ARAM_PIECE_HPP
 
+#include <stdint.h>
 #include <JSystem/JSupport/JSUList.hpp>
 #include <dolphin/ar.h>
 #include <dolphin/os/OSMessage.h>
@@ -10,12 +11,12 @@ class JKRAramBlock;
 class JKRDecompCommand;
 class JKRAMCommand {
 public:
-	typedef void (*AsyncCallback)(u32);
+	typedef void (*AsyncCallback)(uintptr_t);
 
 	JKRAMCommand();
 	~JKRAMCommand();
 
-	u32 getDestination();
+	uintptr_t getDestination();
 
 public:
 	/* 0x00 */ ARQRequest mRequest;
@@ -24,8 +25,8 @@ public:
 
 	/* 0x40 */ s32 mTransferDirection;
 	/* 0x44 */ u32 mDataLength;
-	/* 0x48 */ u32 mSrc;
-	/* 0x4C */ u32 mDst;
+	/* 0x48 */ uintptr_t mSrc;
+	/* 0x4C */ uintptr_t mDst;
 	/* 0x50 */ JKRAramBlock* mAramBlock;
 	/* 0x54 */ u32 field_0x54;
 	/* 0x58 */ AsyncCallback mCallback;
@@ -55,15 +56,17 @@ public:
 	JKRAramPiece();
 	~JKRAramPiece();
 
-	static JKRAMCommand* prepareCommand(int, u32, u32, u32, JKRAramBlock*,
+	static JKRAMCommand* prepareCommand(int, uintptr_t, uintptr_t, u32,
+	                                    JKRAramBlock*,
 	                                    JKRAMCommand::AsyncCallback);
 	static void sendCommand(JKRAMCommand*);
 
-	static JKRAMCommand* orderAsync(int, u32, u32, u32, JKRAramBlock*,
+	static JKRAMCommand* orderAsync(int, uintptr_t, uintptr_t, u32,
+	                                JKRAramBlock*,
 	                                JKRAMCommand::AsyncCallback);
 	static bool sync(JKRAMCommand*, int);
 	static void syncAll(int);
-	static bool orderSync(int, u32, u32, u32, JKRAramBlock*);
+	static bool orderSync(int, uintptr_t, uintptr_t, u32, JKRAramBlock*);
 	static void startDMA(JKRAMCommand*);
 	static void doneDMA(u32);
 
@@ -72,7 +75,8 @@ private:
 	static void unlock() { OSUnlockMutex(&mMutex); }
 };
 
-inline bool JKRAramPcs(int direction, u32 source, u32 destination, u32 length,
+inline bool JKRAramPcs(int direction, uintptr_t source, uintptr_t destination,
+                       u32 length,
                        JKRAramBlock* block)
 {
 	return JKRAramPiece::orderSync(direction, source, destination, length,
