+++ libs/JSystem/src/J3D/J3DGraphLoader/J3DAnmLoader.cpp
@@ -216,11 +216,11 @@ void J3DAnmFullLoader_v15::readAnmVtxColor(const J3DAnmVtxColorFullData* data)
 
 	for (s32 i = 0; i < dst->mAnmTableNum[0]; i++)
 		dst->mAnmVtxColorIndexData[0][i].mpData
-		    = (void*)((s32)indexPtr0
+		    = (void*)((intptr_t)indexPtr0
 		              + (s32)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
 	for (s32 i = 0; i < dst->mAnmTableNum[1]; i++)
 		dst->mAnmVtxColorIndexData[1][i].mpData
-		    = (void*)((s32)indexPtr1
+		    = (void*)((intptr_t)indexPtr1
 		              + (s32)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
 
 	dst->mColorR = JSUConvertOffsetToPtr<u8>(data, (void*)data->mRValuesOffset);
@@ -434,11 +434,11 @@ void J3DAnmKeyLoader_v15::readAnmVtxColor(const J3DAnmVtxColorKeyData* data)
 
 	for (s32 i = 0; i < dst->mAnmTableNum[0]; i++)
 		dst->mAnmVtxColorIndexData[0][i].mpData
-		    = (void*)((s32)indexPtr0
+		    = (void*)((intptr_t)indexPtr0
 		              + (s32)dst->mAnmVtxColorIndexData[0][i].mpData * 2);
 	for (s32 i = 0; i < dst->mAnmTableNum[1]; i++)
 		dst->mAnmVtxColorIndexData[1][i].mpData
-		    = (void*)((s32)indexPtr1
+		    = (void*)((intptr_t)indexPtr1
 		              + (s32)dst->mAnmVtxColorIndexData[1][i].mpData * 2);
 
 	dst->mColorR
